+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
 ## <next>
+* Fix secondary Pane height in horizontal split with calc(100% - pane1size) style
+* Fix Resizer height in horizontal split
+* Fix demo to be full page width/height with fixed size header
+* Upgraded all npm packages
 
 ## 1.0.3
 * Updated npm packages